Extract InputsHelpIcon to a separate component
